requirements.txt
opencv-python
mediapipe
spotipy
requests
python-dotenv
